+++ NEWS.md
@@ -4,6 +4,8 @@
 YAML metadata of (R)Markdown files.
 * Added a LaTeX prettifier. It can be used for Sweave files. It requires the 
 command-line utility `latexindent`, which is included in MikTeX and TeX Live.
+* Added a Julia prettifier. It requires Julia and the Julia package 
+`JuliaFormatter`.
 * Now the package provides some functions to prettify, which return the pretty 
 code. The addins just call these functions and render the pretty code in 
 RStudio.